% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classificationVariableToFactor}
\alias{classificationVariableToFactor}
\title{Make response variable a factor}
\usage{
classificationVariableToFactor(xy, classifiedVarName)
}
\arguments{
\item{xy}{input data}

\item{classifiedVarName}{name of column for response variable}
}
\value{
vector of classification variable observations as factor
}

