% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateRandomDataFrame}
\alias{generateRandomDataFrame}
\title{Generate random data}
\usage{
generateRandomDataFrame(nrow, ncol, binary = TRUE)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{binary}{TRUE if data should be binary (1/0)}
}
\value{
data.frame
}

