% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLearn.R
\name{hcc_accuracy_estimate}
\alias{hcc_accuracy_estimate}
\title{Evaluate the accurate of hcc.}
\usage{
hcc_accuracy_estimate(hcc, markerfile, truelabel)
}
\arguments{
\item{hcc}{A vector of high confidence cells.}

\item{markerfile}{A list contains the marker gene of each cell type.}

\item{truelabel}{A vector of true cell type.}
}
\value{
A vector of accuracy of each cell type.
}
\description{
Evaluate the accurate of hcc.
}
