% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_discriminant_analysis.R
\name{lda.calcPriorClassProbDist}
\alias{lda.calcPriorClassProbDist}
\title{Get prior distribution to be used in Discriminant Analysis

Calculate prior class distributions. If y is observations (yIsSample=TRUE) priors
are the proprtion each class appears in the data. if y is are unique classes,
each class is assumed to occur with equal probability}
\usage{
lda.calcPriorClassProbDist(y, yIsSample = TRUE)
}
\arguments{
\item{y}{factor vector of classes, either observed or unique}

\item{yIsSample}{TRUE if \param{y} is observations of classes}
}
\value{
data.frame of priors for classes
}

