% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_discriminant_analysis.R
\name{lda}
\alias{lda}
\title{Linear Discriminant Analysis Wrapper}
\usage{
lda(xy, config, removeRowColName, removeRowValue, classVariableName,
  priorDistributionIsSample, dataDir, writeIntermediateDataFiles = FALSE,
  attachOutput = FALSE, modelsFirstVarColumnIndex = 4)
}
\arguments{
\item{xy}{data frame of observations}

\item{config}{data frame containing model definitions}

\item{priorDistributionIsSample}{TRUE to use sample frequency to calculate priors}

\item{dataDir}{directory path to write intermediate data files}

\item{writeIntermediateDataFiles}{TRUE to write intermediate data to disk}

\item{attachOutput}{TRUE to attach to global environment,}

\item{modelsFirstVarColumnIndex}{index of the first column which corresponds
to a model variable
mainly for python clients}
}

