% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_discriminant_analysis.R
\name{lda.runLdaAllDataForModels}
\alias{lda.runLdaAllDataForModels}
\title{Run Multiple Discriminant Analysis - All Data}
\usage{
lda.runLdaAllDataForModels(y, x, models, yClassPriors, modelsFirstVarCol = 4)
}
\arguments{
\item{y}{factor vector of classes}

\item{x}{data frame of observations}

\item{models}{data frame of model definitions, rows define variable sets}

\item{yClassPriors}{numeric vector of class priors in order of levels
of \param{y}}
}
\value{
list with 4 data frames - class tabulations, variable means,
discr. functions and variance ratios
}

