% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_discriminant_analysis.R
\name{lda.runLooLdaForModels}
\alias{lda.runLooLdaForModels}
\title{Leave One Out LDA for several models}
\usage{
lda.runLooLdaForModels(y, x, models, yClassPriors, modelsFirstVarCol = 4)
}
\arguments{
\item{y}{factor vector of classes}

\item{x}{data frame of observations}

\item{models}{data frame of model definitions, rows define variable sets}

\item{yClassPriors}{numeric vector of class priors in order of levels
of \param{y}}
}
\value{
list with two data frames - class tabulation and ?posteriors
}

