% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{legacyMatchesExpectedOutput}
\alias{legacyMatchesExpectedOutput}
\title{Check thet files match in two directories}
\usage{
legacyMatchesExpectedOutput(fileName,
  outDirExpected = "/opt/rlearn/tests/data/expected",
  outDir = "/opt/rlearn/tests/data/output", tol = 0.1)
}
\arguments{
\item{outDir}{}
}
\value{
TRUE if file contents are identical
}

