% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLearn.R
\name{LGC}
\alias{LGC}
\title{Get high confidence cells.}
\usage{
LGC(embedding, markerfile, HCC, weight)
}
\arguments{
\item{embedding}{The reduced dimensional data of scRNA-seq.}

\item{markerfile}{A list contains the marker gene of each cell type.}

\item{HCC}{A vector of high confidence cells.}

\item{weight}{A binary value whether to weight high confidence cells.}
}
\value{
A vector of high confidence cells.
}
\description{
Get high confidence cells.
}
