% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_parent_function_call}
\alias{log_parent_function_call}
\title{Log a function call

The name of the function where this function is called will be
logged at debug level}
\usage{
log_parent_function_call()
}

