% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLearn.R
\name{ReadMarker}
\alias{ReadMarker}
\title{Read marker from plain text, each row represent a celltype, the first
column is the tissue, the second column is celltype name, the latter is marker gene.}
\usage{
ReadMarker(con, sep = "\\t", ...)
}
\arguments{
\item{con}{A connection object or a character string.}

\item{sep}{The field separator character.}

\item{...}{The parameter used to readLines.}
}
\value{
Returns a list of marker genes.
}
\description{
Read marker from plain text, each row represent a celltype, the first
column is the tissue, the second column is celltype name, the latter is marker gene.
}
