% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeRowsByColValue}
\alias{removeRowsByColValue}
\title{Filter data}
\usage{
removeRowsByColValue(xy, removeRowValue, removeRowColName)
}
\arguments{
\item{xy}{input data}

\item{removeRowValue}{value for which rows should be removed}

\item{removeRowColName}{column in which \param{removeRowValue} is checked}
}
\value{
filted dataframe
}

