% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLearn.R
\name{rLearn}
\alias{rLearn}
\title{Predict cell type of scATAC-seq dataset by rLearn}
\usage{
rLearn(counts, markers)
}
\arguments{
\item{counts}{A peak-cell count matrix or Seurat object of scATAC-seq data.}

\item{markerfile}{A list contains the marker region of each cell type.}
}
\value{
A vector of predict label.
}
\description{
Predict cell type of scATAC-seq dataset by rLearn
}
