% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectXVariableSubset}
\alias{selectXVariableSubset}
\title{Subset observations to eligible features}
\usage{
selectXVariableSubset(xy, config, xVarSelectCriteria)
}
\arguments{
\item{xy}{input data}

\item{config}{XVARSELV1.csv or XVARSELV.csv}

\item{xVarSelectCriteria}{??? #: TODO ???}
}
\value{
input data subsetted to variables in \param{config}
}

