% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLearn.R
\name{Unlabeled}
\alias{Unlabeled}
\title{Set unlabeled cells.}
\usage{
Unlabeled(F, markerfile, t3)
}
\arguments{
\item{F}{The predict score matrix of each cells (cells * celltype).}

\item{markerfile}{A list contains the marker gene of each cell type.}

\item{t3}{The threshold to determine the unlabeld cells.}
}
\value{
A predict vector with unlabeled cells.
}
\description{
Set unlabeled cells.
}
