% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{vs.calcCorrelationMatrix}
\alias{vs.calcCorrelationMatrix}
\title{Correlation Matrix}
\usage{
vs.calcCorrelationMatrix(x)
}
\arguments{
\item{x}{data.frame of observations}
}
\value{
pairwise correlation matrix, dimnames correspond to x variables
}

