% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{vs.flattenPairwiseCorrelationMatrix}
\alias{vs.flattenPairwiseCorrelationMatrix}
\title{Pretty correlation matrix as data frame}
\usage{
vs.flattenPairwiseCorrelationMatrix(corMat)
}
\arguments{
\item{corMat}{pairwise correlation matrix}
}
\value{
3 column data.frame with pairwise variable correlations
}

