% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{vs.flattenVariableNameSubsets}
\alias{vs.flattenVariableNameSubsets}
\title{Extract variable name subsets from variable selection process

Create a new dataframe with all of the variable names
Inupt array is a 3 dimensional array Where rows are lda solutions from 1 to m number of
variables. Columns are variable names from 1 ("Var.1") to x ("Var.x")}
\usage{
vs.flattenVariableNameSubsets(lviVariableSets, xNames)
}
\arguments{
\item{lviVariableSets}{result of call to improve()}

\item{xNames}{names of X variables used in improve()}
}
\value{
dataframe summarizing models
}

