% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{vs.getVarCorrelations}
\alias{vs.getVarCorrelations}
\title{Calculate variable pairwise correlations}
\usage{
vs.getVarCorrelations(xy, uniqueVars, removeRowValue = -1,
  removeRowColName = "SORTGRP")
}
\arguments{
\item{xy}{observations}

\item{attachOutput}{TRUE to attach to global environment,
mainly for python clients}
}
\value{
data.frame of pairwise variable correlations
}

