% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{vs.improveVariableSelection}
\alias{vs.improveVariableSelection}
\title{Improve routine

Improve model selection under constraints}
\usage{
vs.improveVariableSelection(lviHmat, minNumVar, maxNumVar, nSolutions,
  criteriaVarName)
}
\arguments{
\item{lviHmat}{ouput from ldaHmat}

\item{minNumVar}{minimum number of variables desired for model}

\item{maxNumVar}{maximum number of variables desired for model}

\item{nSolutions}{number of solutions to find for model}
}
\value{
???
}

