% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{vs.removeDegenerateColumns}
\alias{vs.removeDegenerateColumns}
\title{Subset data to columns with information}
\usage{
vs.removeDegenerateColumns(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
\param{x} dataframe with degenerate variables removed
}

