% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{vs.runLDA}
\alias{vs.runLDA}
\title{Run LDA

https://cran.r-project.org/web/packages/subselect/subselect.pdf}
\usage{
vs.runLDA(y, x)
}
\arguments{
\item{y}{response variable as factor}

\item{x}{explanatory variables as data frame}
}
\value{
???
}

