% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{vs.selectVars}
\alias{vs.selectVars}
\title{Select important variables from observations}
\usage{
vs.selectVars(xy, config, yName = "CLPRDP", removeRowValue = -1,
  removeRowColName = "SORTGRP", xVarSelectCriteria = "X",
  improveCriteriaVarName = "xi2", minNumVar = 1, maxNumVar = 10,
  nSolutions = 10)
}
\arguments{
\item{attachOutput}{TRUE to attach to global environment,
mainly for python clients}
}
\value{
dataframe of variables and models as selected by ldaHmat and improve
}

