% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLearn.R
\name{GetHCC}
\alias{GetHCC}
\title{Get high confidence cells.}
\usage{
GetHCC(traindata, markerfile, t1, t2, method = "proxyC")
}
\arguments{
\item{traindata}{The scaled data of scRNA-seq.}

\item{markerfile}{A list contains the marker gene of each cell type.}

\item{t1}{The marker gene expression threshold to determine high confidence cells, default 0.5.}

\item{t2}{The correlation threshold to determine high confidence cells, default 0.8.}

\item{cor_method}{Method to calculate correlation, eg. proxyC and base. Default: proxyC.}
}
\value{
A vector of high confidence cells.
}
\description{
Get high confidence cells.
}
